refactor: 입력한 기능 값에 따라서 기능을 수행하도록 FeatureType 리팩터링
