--- src/main/java/domain/Command.java
@@ -1,6 +1,0 @@
-package domain;
-
-@FunctionalInterface
-public interface Command {
-    void execute();
-}
